fix(gnmi): recover telemetry after TLS rollback
